+++ src/cli/cli-commands/list-scrapers.ts
@@ -26,6 +26,8 @@ export function $initializeListScrapersCommand(program: Program) {
 
     listModelsCommand.action(
         handleActionErrors(async () => {
+            // TODO: !!!!!! Do not allow on REMOTE_SERVER strategy
+
             const scrapers = await $provideScrapersForNode({});
             const executables = await $provideExecutablesForNode();