+++ src/errors/PipelineExecutionError.ts
@@ -23,5 +23,5 @@ export class PipelineExecutionError extends Error {
 }
 
 /**
- * TODO: !!!!!! Add id to all errors
+ * TODO: [🧠][🌂] Add id to all errors
  */