+++ src/errors/WrappedError.ts
@@ -15,17 +15,12 @@ export class WrappedError extends Error {
         console.error(tag, whatWasThrown);
 
         super(
-            spaceTrim(
-                (block) => `
-                    ${/* Fixing tests !!!  block(valueToString(whatWasThrown)) */ block(`non-Error object was thrown`)}
+            spaceTrim(`
+                Non-Error object was thrown
 
-                    Note: Look for ${tag} in the console for more details
-                    !!! Note: \`WrappedError\` indicates that somewhere in the code non-Error object was thrown and it was wrapped
-
-                    Please report issue on ${ADMIN_EMAIL}
-
-                `,
-            ),
+                Note: Look for ${tag} in the console for more details
+                Please report issue on ${ADMIN_EMAIL}
+            `),
         );
         Object.setPrototypeOf(this, WrappedError.prototype);
     }